Rewrite request in pure callbacks
Something is leaking. I suspect that the
Buses, even with the 'air gap' of setTimeout,
are leaking due to their circular nature, but it's
really hard to pin down.

Maybe I should write `request.ls` for XMLHttpRequest
and run this in the browser, where I can at least
track memory more easily.