Ignore catalogs older than latest
In hindsight, this was obvious and a lot easier than the
whole 'regression' system.

Also increased the limiter delay to test the effect on latency.

Looks like past 2000ms, the latency settles to about ~12s average,
which is probably fine for most people. Not as good as the 7s average,
but it'll do, espcially if I want to stream multiple boards with the
same limit.

Also, fixed the `last5` detection for sticky threads, which only have
one last reply in the catalog.