Save state to /tmp periodically
Since resync with 4chan is expensive, having a persistent
state to sync to helps restarts during developement.

Eventually, I want to be able to GET/PUT to
a url e.g. couchdb so I can run fountain in a
heroku-like read-only environment, but this will
do for development.

Probably eventually need to log full events to restore
Last-Event-ID, depending on how I do that.