LiveScript dependency, properly
This means we don't need it installed globally. Doing that is a
versioning nightmare and super awkward to install. (There was a TODO for
it anyway.)

Had to shuffle around the requires a bit; looks like LiveScript has
decided to change stuff since this was last looked at. It's less pretty
this way, but let's make it work first.

Added a prepublish script while I was in there, to automatically do the
compilation step described in the README.